Revert zephyr refactor
