+++ src/DataKML.h
@@ -20,6 +20,9 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include "tinyxml.h"
 #include "com_manager.h"
 
+#ifdef USE_IS_INTERNAL
+#	include "../../cpp/libs/families/imx/IS_internal.h"
+#endif
 
 struct sKmlLogData
 {
