+++ src/ISDataMappings.cpp
@@ -28,7 +28,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include "data_sets.h"
 
 #ifdef USE_IS_INTERNAL
-#include "../is-common/libs/IS_internal.h"
+#include "../../cpp/libs/families/imx/IS_internal.h"
 #endif
 
 using namespace std;
